# docs/swagger.yaml
basePath: /
definitions:
  handlers.APIError:
    properties:
      errors:
        items:
          type: string
        type: array
    type: object
  models.GetFoundingCreatorStatsResponse:
    properties:
      slotsAvailable:
        type: integer
      slotsTaken:
        type: integer
      slotsTotal:
        type: integer
    type: object
  models.GetPreviewRicesResponse:
    properties:
      downloadCount:
        type: integer
      id:
        type: string
      price:
        type: number
      starCount:
        type: integer
      tags:
        items:
          type: string
        type: array
      thumbnailUrl:
        type: string
      title:
        type: string
    type: object
  models.GetWaitlistEmailCountResponse:
    properties:
      count:
        type: integer
    type: object
host: 127.0.0.1:3000
info:
  contact: {}
  description: REST API for RiceHub waitlist frontend.
  title: RiceHub Waitlist API
  version: 1.6.0
paths:
  /founders:
    get:
      description: Returns total, taken, and available founding creator slots
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetFoundingCreatorStatsResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.APIError'
      summary: Get founding creator slot statistics
      tags:
      - founders
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Submit application to become a founding creator
      parameters:
      - description: Username (4-14 alphanumeric characters)
        in: formData
        name: username
        required: true
        type: string
      - description: Email address
        in: formData
        name: email
        required: true
        type: string
      - description: URL to dotfiles repository or r/unixporn post
        in: formData
        name: dotfilesUrl
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/handlers.APIError'
        "409":
          description: Username or email already taken
          schema:
            $ref: '#/definitions/handlers.APIError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.APIError'
      summary: Register as founding creator
      tags:
      - founders
  /rices:
    get:
      description: Returns a list of all preview rices ordered by creation date
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.GetPreviewRicesResponse'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.APIError'
      summary: Get all preview rices
      tags:
      - rices
    post:
      consumes:
      - multipart/form-data
      description: Uploads thumbnail to S3 and stores rice metadata
      parameters:
      - description: Admin secret key
        in: header
        name: X-Admin-Secret
        required: true
        type: string
      - description: Rice title
        in: formData
        name: title
        required: true
        type: string
      - description: Price (e.g. 15.29, must be > 0)
        in: formData
        name: price
        type: number
      - description: Thumbnail image (png, jpeg, webp)
        in: formData
        name: thumbnail
        required: true
        type: file
      - description: Star count
        in: formData
        name: starCount
        required: true
        type: integer
      - description: Download count
        in: formData
        name: downloadCount
        required: true
        type: integer
      - collectionFormat: csv
        description: Tags
        in: formData
        items:
          type: string
        name: tags
        type: array
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/handlers.APIError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handlers.APIError'
        "409":
          description: Title already exists
          schema:
            $ref: '#/definitions/handlers.APIError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.APIError'
      security:
      - AdminSecret: []
      summary: Create a preview rice
      tags:
      - rices
  /waitlist:
    get:
      description: Returns total count
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetWaitlistEmailCountResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.APIError'
      summary: Get amount of users in waitlist
      tags:
      - waitlist
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Adds a new email address to the waitlist
      parameters:
      - description: Email address
        in: formData
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/handlers.APIError'
        "409":
          description: Email already on waitlist
          schema:
            $ref: '#/definitions/handlers.APIError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/handlers.APIError'
      summary: Add email to waitlist
      tags:
      - waitlist
securityDefinitions:
  AdminSecret:
    in: header
    name: X-Admin-Secret
    type: apiKey
swagger: "2.0"